Add files via upload
Application and results of new curricula of arts in the real-time classroom. 
  
This learning tool is designed to assist teachers in assessing the implementation of new curricula during their daily teaching and collecting and editing data. It is an interactive open resource created using the Moodle media tool 'Moodle- lesson' in 2021-2022. The tool is a branching scenario-type tool and was developed as part of a MOOC self-paced teacher training program for the New Visual Studies curriculum. This curriculum is designed for middle and high schools in Greece.

The project (MIS) 5035543 was a collaboration between the Institution of Educational Policy of Greece and the drafting group of expert-art-teacher curricula editors, and their supervisors in 2021. The resource has been integrated into the Institution's Information System on a Moodle site, which can be accessed by enrolling in the course.

It is important to note that this assessment is free and not intended to evaluate mastery of the above course. The digital transfer and transformation from text & power-point to-moodle was carried out by Fotini Kafida, under the supervision of the Committee of the MOOC Educational Projects of the Institution of Educational Policy of Greece. This resource is designed to accompany teachers on their journey of developing and refining their teaching strategies and is a valuable tool for assessing the application of new curricula in actual teaching situations.
________________________________________________________________

«Εφαρμογή και αποτελέσματα των νέων προγραμμάτων σπουδών των τεχνών στην τάξη σε πραγματικό χρόνο». Αυτό είναι ένα  εργαλείο αναστοχασμού, που  έχει σχεδιαστεί για να βοηθήσει τους εκπαιδευτικούς στην αξιολόγηση της εφαρμογής νέων προγραμμάτων σπουδών κατά τη διάρκεια της καθημερινής διδασκαλίας τους και τη συλλογή και επεξεργασία δεδομένων και στη δική τους αναστοχαστική διαδικασία. Είναι ένας διαδραστικός ανοιχτός πόρος που δημιουργήθηκε χρησιμοποιώντας το εργαλείο πολυμέσων Moodle "Moodle- lesson" το 2021-2022. Το εργαλείο moodle είναι τύπος σεναρίου διακλάδωσης (branching scenario) που προσφέρει διαφοροποιημένες διαδρομές και αναπτύχθηκε ως μέρος MOOC, ανοιχτού διαδικτυακού προγράμματος κατάρτισης εκπαιδευτικών για το Νέο πρόγραμμα σπουδών των Εικαστικών του Γυμνασίου. 

Το έργο (MIS) 5035543 ήταν μια συνεργασία μεταξύ του Ιδρύματος Εκπαιδευτικής Πολιτικής της Ελλάδας(ΙΕΠ) και της συντακτικής ομάδας των εμπειρογνωμόνων εκπονητών/ εποπτών για τα Νέα ΠΣ Εικαστικών του Γυμνασίου το 2021. Το περιεχόμενο της δραστηριότητας μετατράπηκε σε δυναμικό-πολυμεσικό-διαδραστικό από το αρχικό στατικό ppt και αναπροσαρμόστηκε, διατηρώντας ωστόσο το περιεχόμενο και κάποιες αρχικές στατικές εικόνες της συγγραφικής ομάδας.  Ο πόρος έχει ενσωματωθεί στο Πληροφοριακό σύστημα του Ιδρύματος σε ιστότοπο Moodle, στον οποίο η πρόσβαση γίνεται  με την εγγραφή στο μάθημα. Είναι σημαντικό να σημειωθεί ότι αφορά προαιρετική δραστηριότητα- βήμα εργασίας στο παραπάνω μάθημα, που ενώ, εκτός από διαφάνειες, δυναμικά στοιχεία .gif, και αφήγηση περιέχει κουίζ, δεν βαθμολογείται ώστε να καθορίζει την ολοκλήρωση του παραπάνω μαθήματος. 
 Ψηφιακή μεταφορά και μετατροπή από κείμενο & στατικό power-point σε διαδραστικό αντικείμενο moodle: Φωτεινή Καφίδα, υπό την επίβλεψη της Επιτροπής Ολοκλήρωσης Έργου του MOOC του Ιδρύματος Εκπαιδευτικής Πολιτικής της Ελλάδας.